requirements.txt
amqp==2.5.0
ariadne==0.5.0
asgiref==3.2.5
atomicwrites==1.3.0
attrs==19.1.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
bleach==3.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
construct==2.5.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.2.3
django-appconf==1.0.3
django-bleach==0.6.1
django-bootstrap4==1.1.1
django-cogwheels==0.3
django-compressor==2.4
django-debug-toolbar==1.11
django-eventtools==0.9.11
django-extensions==2.2.9
django-haystack==2.8.1
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-libsass==0.8
django-machina==1.1
django-modelcluster==5.0.1
django-mptt==0.10.0
django-sass-processor==0.8
django-simple-sso==0.14.0
django-taggit==1.2.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.9.4
draftjs-exporter==2.1.7
Faker==1.0.7
future==0.18.2
geocoder==1.38.1
GraphQL-core-next==1.0.5
html5lib==1.0.1
icalendar==4.0.5
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
kombu==4.6.3
l18n==2018.5
libsass==0.19.4
Markdown==2.6.11
markdown2==2.3.8
Misago==0.26.0
more-itertools==7.1.0
oauthlib==3.0.2
packaging==19.0
pefile==2019.4.18
Pillow==6.1.0
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
pytest-mock==1.10.4
python-dateutil==2.8.0
python-multipart==0.0.5
python-ptrace==0.9.4
python3-openid==3.1.0
pytz==2019.1
ratelim==0.1.6
rcssmin==1.0.6
redis==3.2.1
requests==2.22.0
requests-oauthlib==1.2.0
responses==0.10.6
rjsmin==1.1.0
six==1.12.0
snapshottest==0.5.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.2
sqlparse==0.3.0
starlette==0.12.2
termcolor==1.1.0
text-unidecode==1.2
typing-extensions==3.7.4
Unidecode==1.0.23
urllib3==1.25.3
vine==1.3.0
wagtail==2.8
wagtail-cache==0.5.2
wagtail-import-export==0.2
wagtail-instance-selector==1.2.1
wagtailfontawesome==1.2.1
wagtailmenus==3.0.1
wcwidth==0.1.7
webencodings==0.5.1
webservices==0.7
Willow==1.3
zipp==0.5.2
